+++ src/components/task/existing-task/task-card/index.jsx
@@ -2,12 +2,12 @@ import React, { useState } from "react";
 import propTypes from "prop-types";
 import dayjs from "dayjs";
 
-import styles from "components/task/existing-card/index.module.scss";
+import styles from "components/task/existing-task/task-card/index.module.scss";
 import TextArea from "components/base/text-area";
 import Button from "components/base/button";
 import { TASK_TEXTAREA_NUM_OF_ROW } from "utils/const";
 import { validateDayjsDate } from "utils/helper/validation";
-import ButtonContainer from "./button-container";
+import ButtonContainer from "../button-container";
 
 function TaskCard({ title, createdAt, isCompleted, isTaskOnEditMode }) {
   const [isTextAreaVisible, setIsTextAreaVisible] = useState(isTaskOnEditMode);
